IQTest/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
aiosqlite
pydantic
python-dotenv
httpx
fpdf2
pytest
pytest-asyncio
